OpenAI GPT Virtual Assistant/requirements.txt
openai
google-cloud-texttospeech
PyAudio
SpeechRecognition
pygame